backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5
python-multipart==0.0.9
httpx==0.27.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
google-generativeai==0.7.2
groq==0.9.0
splitwise==2.3.0
python-telegram-bot==20.8
apscheduler==3.10.4
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
Pillow==10.3.0
PyPDF2==3.0.1
aiofiles==23.2.1
